requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart

# OCR
pytesseract
pillow

# LLM
transformers
accelerate
safetensors
torch